% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{caseCount}
\alias{caseCount}
\title{Count the number of cases in an experiment}
\usage{
caseCount(exp, pattern = "*")
}
\arguments{
\item{pattern}{Case name filter}
}
\description{
Count the number of cases that match the given pattern. Useful to check before running a deletion
}
