% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{clearJobErrors}
\alias{clearJobErrors}
\title{Clear all job errors}
\usage{
clearJobErrors(queue = "")
}
\arguments{
\item{queue}{The name of the queue to which the jobs will be submitted.}
}
\description{
Clear all job errors
}
