% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis.R
\name{clientCount}
\alias{clientCount}
\title{Get the number of clients connected to the Redis server (including non-workers)}
\usage{
clientCount()
}
\description{
Get the number of clients connected to the Redis server (including non-workers)
}
