% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{deleteCases}
\alias{deleteCases}
\title{Delete cases of experiment according to a pattern}
\usage{
deleteCases(exp, pattern, sim = TRUE, fixed = FALSE)
}
\arguments{
\item{pattern}{Case name filter}

\item{sim}{Whether to simulate the deletion, showing the number of cases that would have been deleted.}

\item{fixed}{Whether to match the given pattern as-is}
}
\description{
Delete cases from the experiment that match the given name pattern
}
