% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment_computeResultWithData}
\alias{experiment_computeResultWithData}
\title{Evaluate result with case data}
\usage{
experiment_computeResultWithData(
  name,
  resultfun,
  verbose = 1,
  dataArgs = c("data", "dataseed", "numgroups", "re", "numobs", "numtraj", "noise",
    "propnoise", "int.sd", "slo.sd", "q.sd", "props")
)
}
\arguments{
\item{resultFun}{A function with arguments `(tsdata, output)`}
}
\description{
Evaluate the result with access to the case-related dataset
}
