% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment_getOutputTableOf}
\alias{experiment_getOutputTableOf}
\title{Get post-processed output table}
\usage{
experiment_getOutputTableOf(name, outfun)
}
\arguments{
\item{outfun}{optional function that is called on the output of each field, for returning additional scalar columns}
}
\description{
Get post-processed output table
}
