% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseframe.R
\name{generateCaseNames}
\alias{generateCaseNames}
\title{Generate case names for the given case settings}
\usage{
generateCaseNames(cases, columns = names(cases))
}
\arguments{
\item{cases}{A \code{data.frame} containing different parameter settings per row. A "data" and "method" column are required, containing a character or factor denoting the name of the variable to evaluate.}
}
\description{
Generate case names for the given case settings
}
