% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{getCases}
\alias{getCases}
\title{Get the evaluated case settings in a data.frame}
\usage{
getCases(exp, ...)
}
\arguments{
\item{exp}{Name of the experiment}

\item{...}{
  Arguments passed on to \code{\link[=getExperimentResults]{getExperimentResults}}
  \describe{
    \item{\code{subset}}{Optional logical expression or case-name regex pattern for filtering cases}
    \item{\code{include}}{Whether to include matching cases in the output, or exclude}
  }}
}
\description{
Get the evaluated case settings in a data.frame
}
