% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{getExperimentOutputs}
\alias{getExperimentOutputs}
\title{Get a list of all method outputs per case}
\usage{
getExperimentOutputs(exp, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=getExperimentResults]{getExperimentResults}}
  \describe{
    \item{\code{subset}}{Optional logical expression or case-name regex pattern for filtering cases}
    \item{\code{include}}{Whether to include matching cases in the output, or exclude}
  }}
}
\description{
Get a list of all method outputs per case
}
