% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{getExperimentResults}
\alias{getExperimentResults}
\title{Get the raw experiment result structure}
\usage{
getExperimentResults(exp, subset = "*", include = TRUE, ...)
}
\arguments{
\item{subset}{Optional logical expression or case-name regex pattern for filtering cases}

\item{include}{Whether to include matching cases in the output, or exclude}

\item{...}{Ignored.}
}
\value{
A list of the raw experiment result structures of the matching cases
}
\description{
Get the raw experiment result structure
}
