% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{isCaseEvaluated}
\alias{isCaseEvaluated}
\title{Check whether the given cases have been evaluated}
\usage{
isCaseEvaluated(exp, caseNames)
}
\arguments{
\item{caseNames}{Character vector of case names}
}
\description{
Check whether the given cases have been evaluated
}
