% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis.R
\name{isConnected}
\alias{isConnected}
\title{Assess whether the session has a connection with the Redis server}
\usage{
isConnected()
}
\description{
Assess whether the session has a connection with the Redis server
}
