% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{isJobQueued}
\alias{isJobQueued}
\title{Test whether a the given job has already been queued}
\usage{
isJobQueued(exp, caseNames, queue = "")
}
\description{
Test whether a the given job has already been queued
}
