% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{jobMonitor}
\alias{jobMonitor}
\alias{jobs}
\title{Monitor job progress}
\usage{
jobMonitor(queue = "")
}
\arguments{
\item{queue}{The name of the queue to which the jobs will be submitted.}
}
\description{
Monitor job progress
}
