% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mergeExperiment}
\alias{mergeExperiment}
\title{Merge an experiment into another}
\usage{
mergeExperiment(exp, exp2, pattern = "*")
}
\arguments{
\item{exp2}{The experiment to merge into \code{exp}}

\item{pattern}{Case name filter}
}
\description{
Merge an experiment into another
}
