% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{simworkr-package}
\alias{simworkr-package}
\alias{_PACKAGE}
\title{simworkr: Distributed simulation framework}
\description{
The simworkr package provides a distributed simulation framework
centered around a Redis database to which worker processes connect to obtain new jobs to compute.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/niekdt/simworkr}
  \item Report bugs at \url{https://github.com/niekdt/simworkr/issues}
}

}
\author{
\strong{Maintainer}: Niek Den Teuling \email{niekdenteuling@gmail.com} (\href{https://orcid.org/0000-0003-1026-5080}{ORCID})

}
