% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simworkr.R
\name{simworkr}
\alias{simworkr}
\title{Run a simulation worker}
\usage{
simworkr(
  host = getOption("redis.host", "localhost"),
  port = getOption("redis.port", 6379),
  password = getOption("redis.pwd"),
  name = NULL,
  queue = "",
  maxJobs = Inf,
  catchError = FALSE,
  envir = parent.frame()
)
}
\arguments{
\item{host}{The Redis server host name or inet address (optional, character). The default value is "localhost".}

\item{port}{The Redis port number (optional, numeric or integer). The default value is 6379L.}

\item{password}{Redis authentication password.}

\item{name}{Worker name. If not specified, a name is generated.}

\item{maxJobs}{The maximum number of jobs that this worker is allowed to fetch from the queue.}

\item{catchError}{Whether to catch computation errors and continue.}

\item{envir}{The environment from which to retrieve the data and method. Note that the data and method calls are evaluated in isolation.}
}
\description{
Run a simulation worker
}
