% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{submitJobs}
\alias{submitJobs}
\title{Submit a series of cases as jobs}
\usage{
submitJobs(exp, cases, inOrder = FALSE, renew = FALSE, queue = "")
}
\arguments{
\item{exp}{Name of the experiment to which the cases belong.}

\item{cases}{A \code{data.frame} containing different parameter settings per row. A "data" and "method" column are required, containing a character or factor denoting the name of the variable to evaluate.}

\item{inOrder}{Whether to submit the cases in order. Shuffling the order of the cases results in more reliable time estimates.}

\item{renew}{Whether to redo cases that have already been evaluated.}

\item{queue}{The name of the queue to which the jobs will be submitted.}
}
\description{
Submit a series of cases as jobs
}
